requirements.txt
asgiref==3.10.0
cloudinary==1.41.0
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
